+++ mobile/lib/features/voice_hub/presentation/widgets/voice_room/voice_room_connection_overlays.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../providers/chat_room_providers.dart';
-import '../providers/room_fragment_providers.dart';
-import '../providers/voice_room_mention_notice_provider.dart';
+import '../../providers/chat_room_providers.dart';
+import '../../providers/room_fragment_providers.dart';
+import '../../providers/voice_room_mention_notice_provider.dart';
 import 'voice_room_mention_notice_banner.dart';
 import 'voice_room_reconnect_banner.dart';